JBPM-9385 : Stunner - Text area for scripts is cropped/shifted - Fixed Bugs

+++ kie-wb-common-stunner/kie-wb-common-stunner-sets/kie-wb-common-stunner-bpmn/kie-wb-common-stunner-bpmn-api/src/main/java/org/kie/workbench/common/stunner/bpmn/definition/property/diagram/AbstractDataTypeCache.java
@@ -191,7 +191,7 @@ public abstract class AbstractDataTypeCache {
         allDataTypes.remove("String");
         allDataTypes.remove("Integer");
         allDataTypes.remove("Boolean");
-        allDataTypes.remove("Double");
+        allDataTypes.remove("Float");
         return allDataTypes;
     }
 }
